Read stdin for dayplan data.
In addition to the -f flag, it is now possible to let the dayplan
command line utility read dayplan data from stdin.

Fixes #32.

+++ test/system/dpltest.py
@@ -4,23 +4,36 @@ import subprocess
 import tempfile
 
 
-def dpl_run(cmd, content, valgrind=False):
-    d = tempfile.mkdtemp()
-    dplfile = os.path.join(d, 'input.dat')
-    with open(dplfile, 'w') as f:
-        f.write(content)
-    try:
-        cmd = [ 'dayplan', '-f', dplfile ] + cmd
-        if valgrind:
-            cmd.insert(0, 'valgrind')
-        p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+def dpl_run(cmd, content, valgrind=False, stdin=False):
+    cmd = [ 'dayplan' ] + cmd
+    if valgrind:
+        cmd.insert(0, 'valgrind')
+    if stdin:
+        p = subprocess.Popen(
+                cmd, 
+                stdout=subprocess.PIPE, 
+                stderr=subprocess.PIPE,
+                stdin=subprocess.PIPE)
+        p.stdin.write(content.encode('utf-8'))
         out, err = p.communicate()
-        out, err = out.decode('utf-8'), err.decode('utf-8')
-        if valgrind:
-            assert 'All heap blocks were freed -- no leaks are possible' in err
-        return p.returncode, out, err
-    finally:
-        shutil.rmtree(d)
+    else:
+        d = tempfile.mkdtemp()
+        dplfile = os.path.join(d, 'input.dat')
+        try:
+            with open(dplfile, 'w') as f:
+                f.write(content)
+            cmd.extend([ '-f', dplfile ])
+            p = subprocess.Popen(
+                    cmd, 
+                    stdout=subprocess.PIPE, 
+                    stderr=subprocess.PIPE)
+            out, err = p.communicate()
+        finally:
+            shutil.rmtree(d)
+    out, err = out.decode('utf-8'), err.decode('utf-8')
+    if valgrind:
+        assert 'All heap blocks were freed -- no leaks are possible' in err
+    return p.returncode, out, err
 
 
 def assert_dpl_cmd_error(cmd, content):
@@ -30,10 +43,12 @@ def assert_dpl_cmd_error(cmd, content):
 
 
 def assert_dpl_cmd_ok(cmd, content, expected):
-    ret, out, err = dpl_run(cmd, content, valgrind=False)
+    ret, out, err = dpl_run(cmd, content, valgrind=False, stdin=False)
     print(ret, "\n", out, "\n", err, "\n", expected)
-    #if out.strip() != expected.strip():
-    #   import pdb
-    #   pdb.set_trace()
     assert out.strip() == expected.strip()
     assert ret == 0
+    # ensure the same result when using stdin for input
+    reti, outi, erri = dpl_run(cmd, content, valgrind=False, stdin=True)
+    assert reti == ret
+    assert outi == out
+    assert erri == err
